resolution3: move avatar stuff into its own file
